+++ cellreaction_data.cpp
@@ -1,9 +1,9 @@
 #include "cellreaction_data.h"
 
-CellReactionData::CellReactionData(double k, int prevState, int nextState) :
+CellReactionData::CellReactionData(long double k, int prevState, int nextState) :
     ReactionData<CellData>(k, prevState, nextState) {}
 
-double CellReactionData::rate(const CellData *site) const {
+long double CellReactionData::rate(const CellData *site) const {
     return (site->value() == prevState()) ? rateValue(site) : 0;
 }
