+++ dimerreactiondualstabilization_data.cpp
@@ -1,9 +1,9 @@
 #include "dimerreactiondualstabilization_data.h"
 
-DimerReactionDualStabilizationData::DimerReactionDualStabilizationData(double k, int firstPrevState, int secondPrevState, int nextState) :
+DimerReactionDualStabilizationData::DimerReactionDualStabilizationData(long double k, int firstPrevState, int secondPrevState, int nextState) :
     ReactionData<DimerData>(k, firstPrevState, nextState), _secondPrevState(secondPrevState) {}
 
-double DimerReactionDualStabilizationData::rate(const DimerData *dimer) const {
+long double DimerReactionDualStabilizationData::rate(const DimerData *dimer) const {
     return ((dimer->first->value() == prevState() && dimer->second->value() == _secondPrevState) ||
             (dimer->first->value() == _secondPrevState && dimer->second->value() == prevState())) ?
                 rateValue(dimer) : 0;
