+++ dimerreactionstabilization_data.h
@@ -7,9 +7,9 @@
 class DimerReactionStabilizationData : public ReactionData<DimerData>
 {
 public:
-    DimerReactionStabilizationData(double k, int prevState, int nextState);
+    DimerReactionStabilizationData(long double k, int prevState, int nextState);
 
-    double rate(const DimerData *dimer) const;
+    long double rate(const DimerData *dimer) const;
     void doIt(DimerData *const dimer) const;
 };
