+++ lateralreaction_data.cpp
@@ -1,9 +1,9 @@
 #include "lateralreaction_data.h"
 
-LateralReactionData::LateralReactionData(const RTReactorContext *reactor, double Ea, double epss[], int numOfSpecs) :
+LateralReactionData::LateralReactionData(const RTReactorContext *reactor, long double Ea, long double epss[], int numOfSpecs) :
     ReactionActivationData(reactor, Ea), _numOfSpecs(numOfSpecs)
 {
-    _epss = new double[_numOfSpecs];
+    _epss = new long double[_numOfSpecs];
     for (int i = 0; i < _numOfSpecs; ++i) _epss[i] = epss[i];
 }
 
@@ -11,8 +11,8 @@ LateralReactionData::~LateralReactionData() {
     delete [] _epss;
 }
 
-double LateralReactionData::expFactor(const LateralSiteData *site) const {
-    double lateralEnergy = 0;
+long double LateralReactionData::expFactor(const LateralSiteData *site) const {
+    long double lateralEnergy = 0;
     for (int i = 0; i < _numOfSpecs; ++i) {
         lateralEnergy += site->numOfSpec(i) * _epss[i];
     }
